# api/openapi.yaml
openapi: 3.0.3
info:
  title: DonoEngine
  description: Go-based comic catalog backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: /api/v0
paths:
...
  /categories:
    get:
      tags:
        - Category
      summary: List category.
      operationId: listCategory
      parameters:
        - name: page
          in: query
          description: Page number of results.
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of results.
          schema:
            type: integer
        - name: order_by
          in: query
          description: Sort results returned.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Category list.
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: The total count of category with current filter.
            X-Pagination-Limit:
              schema:
                type: integer
              description: The last page number of category with current filter and limit.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Category
      summary: Add category.
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCategory'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewCategory'
        required: true
      responses:
        '201':
          description: Category added.
          headers:
            Location:
              description: The path of new category.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /categories/{typeID}-{code}:
    get:
      tags:
        - Category
      summary: Get category.
      operationId: getCategory
      parameters:
        - name: typeID
          in: path
          description: Type ID of category type.
          required: true
          schema:
            type: integer
            x-go-type: uint
        - name: code
          in: path
          description: Code of category to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Category
      summary: Update category.
      operationId: updateCategory
      parameters:
        - name: typeID
          in: path
          description: Type ID of category type.
          required: true
          schema:
            type: integer
            x-go-type: uint
        - name: code
          in: path
          description: Code of category to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCategory'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetCategory'
        required: true
      responses:
        '200':
          description: Category updated.
          headers:
            Location:
              description: The path of updated category.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '204':
          description: Category unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Category
      summary: Delete category.
      operationId: deleteCategory
      parameters:
        - name: typeID
          in: path
          description: Type ID of category type.
          required: true
          schema:
            type: integer
            x-go-type: uint
        - name: code
          in: path
          description: Code of category to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Category deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /categories/{typeID}-{code}/relations:
    post:
      tags:
        - Category
      summary: Add category relation.
      operationId: addCategoryRelation
      parameters:
        - name: typeID
          in: path
          description: Type ID of category type.
          required: true
          schema:
            type: integer
            x-go-type: uint
        - name: code
          in: path
          description: Code of category.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCategoryRelation'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewCategoryRelation'
        required: true
      responses:
        '201':
          description: Category relation added.
          headers:
            Location:
              description: The path of new category relation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRelation'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /categories/{typeID}-{code}/relations/{categoryCode}:
    get:
      tags:
        - Category
      summary: Get category relation.
      operationId: getCategoryRelation
      parameters:
        - name: typeID
          in: path
          description: Type ID of category type.
          required: true
          schema:
            type: integer
            x-go-type: uint
        - name: code
          in: path
          description: Code of category.
          required: true
          schema:
            type: string
        - name: categoryCode
          in: path
          description: Code of related category.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category relation gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRelation'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Category
      summary: Update category relation.
      operationId: updateCategoryRelation
      parameters:
        - name: typeID
          in: path
          description: Type ID of category type.
          required: true
          schema:
            type: integer
            x-go-type: uint
        - name: code
          in: path
          description: Code of category.
          required: true
          schema:
            type: string
        - name: categoryCode
          in: path
          description: Code of related category.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCategoryRelation'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetCategoryRelation'
        required: true
      responses:
        '200':
          description: Category relation updated.
          headers:
            Location:
              description: The path of updated category relation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRelation'
        '204':
          description: Category relation unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Category
      summary: Delete category relation.
      operationId: deleteCategoryRelation
      parameters:
        - name: typeID
          in: path
          description: Type ID of category type.
          required: true
          schema:
            type: integer
            x-go-type: uint
        - name: code
          in: path
          description: Code of category.
          required: true
          schema:
            type: string
        - name: categoryCode
          in: path
          description: Code of related category.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Category relation deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
...
components:
  schemas:
    Object:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-go-type: uint
          x-go-name: ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - createdAt
...
    Category:
      type: object
      allOf:
        - $ref: '#/components/schemas/Object'
        - type: object
          properties:
            typeID:
              type: integer
              x-go-type: uint
            typeCode:
              type: string
            code:
              type: string
            name:
              type: string
            relations:
              type: array
              items:
                $ref: '#/components/schemas/CategoryRelation'
          required:
            - typeID
            - typeCode
            - code
            - name
    NewCategory:
      type: object
      properties:
        typeID:
          type: integer
          nullable: true
          x-go-type: uint
        typeCode:
          type: string
          nullable: true
        code:
          type: string
        name:
          type: string
      required:
        - code
        - name
    SetCategory:
      type: object
      properties:
        typeID:
          type: integer
          nullable: true
          x-go-type: uint
        typeCode:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
    CategoryRelation:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        categoryID:
          type: integer
          x-go-type: uint
        categoryCode:
          type: string
      required:
        - createdAt
        - categoryID
        - categoryCode
    NewCategoryRelation:
      type: object
      properties:
        categoryID:
          type: integer
          nullable: true
          x-go-type: uint
        categoryCode:
          type: string
          nullable: true
    SetCategoryRelation:
      type: object
      properties:
        categoryID:
          type: integer
          nullable: true
          x-go-type: uint
        categoryCode:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties: 
            status:
              type: string
            message:
              type: string
          required:
            - status
            - message
      required:
        - error
  responses:
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT